# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  dto.ImportCsvAPIResponse:
    properties:
      data:
        $ref: '#/definitions/dto.ImportCsvResult'
      message:
        type: string
      success:
        type: boolean
    type: object
  dto.ImportCsvResult:
    properties:
      channel_id:
        type: integer
      channel_update_failed:
        type: boolean
      completion_ratio_skipped:
        type: integer
      errors:
        items:
          $ref: '#/definitions/dto.ImportError'
        type: array
      intro_updated:
        type: integer
      models_failed:
        type: integer
      models_imported:
        type: integer
      models_in_csv:
        type: integer
      models_recognized:
        type: integer
      models_skipped:
        type: integer
      models_skipped_no_input:
        type: integer
      models_skipped_non_token:
        type: integer
      persisted_ratio_models:
        items:
          type: string
        type: array
      price_skipped:
        type: integer
      price_updated:
        type: integer
      ratio_persisted:
        items:
          type: string
        type: array
      warnings:
        items:
          $ref: '#/definitions/dto.ImportWarning'
        type: array
    type: object
  dto.ImportError:
    properties:
      model:
        type: string
      reason:
        type: string
    type: object
  dto.ImportWarning:
    properties:
      model:
        type: string
      reason:
        type: string
    type: object
...
info:
  contact: {}
  description: new-api AI 网关 —— 用户与账号相关接口(对外第三方)
  title: new-api 用户 API
  version: "1.0"
paths:
  /channel/{id}/import_models_csv:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 渠道 ID
        in: path
        name: id
        required: true
        type: integer
      - description: UCloud CSV 文件(UTF-8)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: data 为 dto.ImportCsvResult 结构化结果
          schema:
            $ref: '#/definitions/dto.ImportCsvAPIResponse'
        "400":
          description: invalid channel id / no_file / bad_header / invalid_csv_row
            / too_many_rows / bad_encoding / channel_not_found
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "413":
          description: file_too_large
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: CSV 批量导入渠道模型/价格/介绍
      tags:
      - 渠道-模型导入
...
  /user/checkin:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: 获取签到状态
      tags:
      - 用户-其它
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: 每日签到
      tags:
      - 用户-其它
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"